requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-RESTful==0.3.10
flask-restful-swagger==0.20.2
flask-restful-swagger-3==0.5.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-Testing==0.8.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.2
packaging==24.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
